example/re/few-shot/requirements.txt
torch==1.10.0
transformers==4.26.0
hydra-core==1.3.1
wandb==0.13.9